+++ catalog/app/components/Code/index.ts
@@ -1,0 +1,1 @@
+export { default } from './Code'
